Let macros define other macros (#106)
If a macro that doesn't exist is called, the program checks if there are
any definitions left. If there are, it doesn't error out and continues
until either the macro exists, or there are no more definitions left,
causing an error.
Macro definition tokens are removed from parse_macros() to stop the new
check from running forever, causing a stack overflow.