Update changelog (#121)
* Update CHANGELOG.md changes lists

Updates the lists for anything other than added to use words like 'Fix'
or 'Update' to avoid confusion.

* Add hyperlinks to bottom of file

The reason for the Keep a Changelog versions being in square brackets
was to add links at the end of the file, making the lines easier to read
when looking at the raw markdown. This adds those missing links to the
bottom of the file.

* Reverse order of links at bottom of file

Reverses the order of the version links at the bottom of the file. This
only changes how the raw markdown looks, but it'll make sure that the
order of the links matches the order of the versions.