Fix `sbop` keyword not working (#125)
* Fix `sbop` keyword not working

Fixes the `sbop` keyword for scoreboard operations not being replaced
properly. The original way to replace it was actually a very weird way
to go about it, so I'm not sure why it was added that way in the first
place. Wonder who did that?

* Remove ScoreboardOperation token

Gets rid of the scoreboard operation token and any references to it.

* Update CHANGELOG.md