Add line and column numbers to error messages (#128)
* Add line and column numbers to some error messages

Adds line and column numbers to error messages from tokenize.rs via a
new function make_syntax_error.

* Add file to error messages

Adds the canonical filepath as well as the line/column numbers to the
error message (eg. `/path/to/file:5:10`).

* Update invalid tag message

Changes the error message for a misplaced tag to a panic, since tokens
outside functions should be ignored.

* Add line/column numbers to preprocess.rs error

Adds line and column numbers to the macro error from preprocess.rs. This
was done by adding two usizes to the MacroName token. This might not be
the most efficient way to do it, but it was the fastest.

* Fix incorrect error column numbers

Fixes the column number being 1 higher than it should be due to the
newline character incrementing the column integer.

* Update CHANGELOG.md