More public functions and more clear file structure (#132)
* Move code to order files to new function

Moves the code from main.rs used to place global macro files before
others to a new function prioritize_macro_files in files.rs.

* Add new types TagMap and GlobalMacros

Adds a new module types with the types TagMap and GlobalMacros to avoid
confusion when creating the tag map and global macros variables needed
by the compiler.

* Remove unused import from files.rs

* Move code reading vars.toml to new function

Moves the code that reads the vars.toml file to a HashMap to a new
function read_vars_toml in files.rs.

* Update docs in files.rs

* Update tag_map and global_macros definitions

Removes the implied type from the variable definitions of type TagMap
and GlobalMacros.

* Move code to get CLI matches to new function

Moves the code used to get CLI matches from main.rs to a new function in
cli.rs.

* Move code to compile global macros to new function

Moves the code used to tell if a file contains global macros and to pass
different arguments to Compiler::compile to a new function in the
Compiler struct.

* Move `use` in cli.rs to below license comment

* Update CLI file imports

Changes the CLI files to use `use databind` instead of `use crate` or
directly using modules for library functions. This makes sure that the
CLI is using the same functions that another developer would have access
to.

* Change main.rs to more compact `use`

* Update CHANGELOG.md

* Move CLI files to new directory

Moves the files related to the CLI to a separate subdirectory of src/.
This should help make it more obvious which files make up the CLI at a
glance of the file structure.

* Update PR Labeler config

Updates the PR Labeler config to use the new CLI directory for relevant
labels.

* Move library files to new directory

Moves the files related to the library to a separate subdirectory of src/.
This should help make it more obvious which files make up the library at
a glance of the file structure.

* Update CHANGELOG.md

* Update PR Labeler config

Updates the PR Labeler config to use the new library directory for
relevant labels.