Fix get_namespace tag problems
CLoses #140

Fixes a problem where `files::get_namespace` wouldn't work properly if
the path contained the word `functions` at any point. This is actually
not backwards-compatible due to the change of return type from
`Result<&str, &str>` to `Result<String, &str>`, but the API isn't stable
and isn't used by anyone else so it seems ok to release as a patch.