Make test_macro_defs_in_defs passing
This took some messing with the tokenization and the adding of new
functions to the Node enum, but after a bunch of weird code that I
probably won't be able to understand tomorrow, it works! The changes
made to parse_macros should be added to replace_if_and_while as well.