Fix some magic with macros or something
It's been a while since I worked on this project, so I can't even
remember what I was trying to fix when I stopped halfway through the
changes for this commit. Whatever I was doing, it seems to work now. No
additional tests are passing, but no extras are failing either, which
I'll take. There are a lot of debug prints left over, but I'll get rid
of those soon(TM).